rename app-constants to constants
