--- apps/express-server/src/app-constants/index.ts
@@ -1,3 +1,0 @@
-export * from './config';
-export * from './env_vars';
-
