+++ apps/express-server/src/instrument.ts
@@ -1,6 +1,6 @@
 import * as Sentry from '@sentry/node';
 import { nodeProfilingIntegration } from '@sentry/profiling-node';
-import { ENV_VARS } from '@/app-constants';
+import { ENV_VARS } from '@/constants';
 
 Sentry.init({
   dsn: ENV_VARS.sentryDSN,
